refactor: separate TCPServer into TCPServer and RoomManager classes
